:fire: delete unnecessary code
