+++ demo/demo/urls/api.py
@@ -3,7 +3,6 @@ from django.urls import path
 from demo.views import (
     GCPAggregateCandleView,
     GCPAggregateTradeDataView,
-    GCPInferenceView,
 )
 
 urlpatterns = [
@@ -15,5 +14,4 @@ urlpatterns = [
     path(
         "aggregate-candles/", GCPAggregateCandleView.as_view(), name="aggregate_candles"
     ),
-    path("inference/", GCPInferenceView.as_view(), name="inference"),
 ]
