+++ quant_tick/models/candle_types/adaptive_candles.py
@@ -1,13 +1,18 @@
 from datetime import datetime
 from decimal import Decimal
 
+import pandas as pd
+from django.db.models import QuerySet
+
+from quant_tick.constants import Frequency
+from quant_tick.lib import get_existing, get_min_time
 from quant_tick.utils import gettext_lazy as _
 
-from .adaptive_mixin import AdaptiveMixin
+from ..trades import TradeData
 from .constant_candles import ConstantCandle
 
 
-class AdaptiveCandle(AdaptiveMixin, ConstantCandle):
+class AdaptiveCandle(ConstantCandle):
     """Dynamic-threshold bars that adapt to recent average activity levels.
 
     Instead of using a fixed threshold like constant bars, adaptive bars adjust their
@@ -50,6 +55,15 @@ class AdaptiveCandle(AdaptiveMixin, ConstantCandle):
             data["target_value"] = self.get_moving_average_value(timestamp)
         return data
 
+    def get_trade_data_for_moving_average(self, timestamp: datetime) -> QuerySet:
+        """Get trade data for moving average."""
+        days = self.json_data["moving_average_number_of_days"]
+        delta = pd.Timedelta(f"{days}d")
+        ts = get_min_time(timestamp, value="1d") - delta
+        return TradeData.objects.filter(
+            symbol=self.symbol, timestamp__gte=ts, timestamp__lt=ts + delta
+        )
+
     def get_moving_average_value(self, timestamp: datetime) -> Decimal:
         """Get moving average value."""
         days = self.json_data["moving_average_number_of_days"]
@@ -68,6 +82,16 @@ class AdaptiveCandle(AdaptiveMixin, ConstantCandle):
             ]
         )
         return total / days / self.json_data["target_candles_per_day"]
+
+    def can_aggregate(self, timestamp_from: datetime, timestamp_to: datetime) -> bool:
+        """Can aggregate."""
+        can_agg = super().can_aggregate(timestamp_from, timestamp_to)
+        if can_agg:
+            trade_data = self.get_trade_data_for_moving_average(timestamp_from)
+            existing = get_existing(trade_data.values("timestamp", "frequency"))
+            days = self.json_data["moving_average_number_of_days"]
+            return len(existing) >= Frequency.DAY * days
+        return False
 
     def should_aggregate_candle(self, data: dict) -> bool:
         """Should aggregate candle."""
