+++ quant_tick/serializers/__init__.py
@@ -1,11 +1,10 @@
-from .candles import CandleDataSerializer, CandleSerializer, StrategySerializer
+from .candles import CandleDataSerializer, CandleSerializer
 from .timeago import TimeAgoSerializer, TimeAgoWithRetrySerializer
 from .timeframe import TimeFrameSerializer, TimeFrameWithLimitSerializer
 
 __all__ = [
     "CandleDataSerializer",
     "CandleSerializer",
-    "StrategySerializer",
     "TimeFrameSerializer",
     "TimeFrameWithLimitSerializer",
     "TimeAgoSerializer",
