+++ quant_tick/serializers/candles.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from quant_tick.models import Candle, Strategy
+from quant_tick.models import Candle
 
 
 class CandleSerializer(serializers.ModelSerializer):
@@ -8,14 +8,6 @@ class CandleSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Candle
-        fields = ("code_name",)
-
-
-class StrategySerializer(serializers.ModelSerializer):
-    """Strategy serializer."""
-
-    class Meta:
-        model = Strategy
         fields = ("code_name",)
