--- quant_tick/tests/models/candle_types/test_cluster_bucket_mixin.py
@@ -1,493 +1,0 @@
-from decimal import Decimal
-
-import pandas as pd
-from django.test import TestCase
-
-from quant_tick.constants import FileData
-from quant_tick.lib import WARMUP_MIN_CLUSTERS
-from quant_tick.models import ConstantCandle, TimeBasedCandle
-
-
-def make_trade(
-    timestamp: pd.Timestamp,
-    price: str,
-    volume: str,
-    notional: str,
-    tick_rule: int,
-) -> dict:
-    """Build a trade row for clustering."""
-    return {
-        "uid": f"t{timestamp}",
-        "timestamp": timestamp,
-        "nanoseconds": 0,
-        "price": Decimal(price),
-        "volume": Decimal(volume),
-        "notional": Decimal(notional),
-        "ticks": 1,
-        "tickRule": tick_rule,
-    }
-
-
-def warmed_cache() -> dict:
-    """Return cache with EMA already warmed up.
-
-    Pre-seeds cluster_ema_count to skip warmup for tests verifying bucket stats.
-    """
-    return {
-        "cluster_ema_count": WARMUP_MIN_CLUSTERS,
-        "cluster_ema_mean": 7.0,  # ~log(1000)
-        "cluster_ema_var": 1.0,
-    }
-
-
-class TimeBasedClusterBucketTest(TestCase):
-    """Test ClusterBucketMixin with TimeBasedCandle."""
-
-    def _make_candle(self, **overrides: object) -> TimeBasedCandle:
-        """Create candle with cluster bucket stats enabled."""
-        json_data = {
-            "source_data": FileData.FILTERED,
-            "window": "1h",
-            "cluster_bucket_stats": True,
-        }
-        json_data.update(overrides)
-        return TimeBasedCandle(json_data=json_data)
-
-    def test_emits_at_time_intervals(self) -> None:
-        """Candles emitted at window boundaries."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("30min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("90min"), "101", "1000", "101000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("2h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 2)
-        self.assertEqual(data[0]["timestamp"], now)
-        self.assertEqual(data[1]["timestamp"], now + pd.Timedelta("1h"))
-
-    def test_percentile_buckets_present(self) -> None:
-        """All percentile buckets present in output."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("20min"), "101", "2000", "202000", -1),
-            make_trade(now + pd.Timedelta("30min"), "99", "3000", "297000", 1),
-            make_trade(now + pd.Timedelta("40min"), "98", "4000", "392000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 1)
-        for p in [0, 25, 50, 75, 90, 95, 99]:
-            key = f"p{p}"
-            self.assertIn(key, data[0])
-            self.assertIn("buyClusterCount", data[0][key])
-            self.assertIn("totalClusterCount", data[0][key])
-            self.assertIn("buyClusterNotional", data[0][key])
-            self.assertIn("totalClusterNotional", data[0][key])
-
-    def test_total_cluster_count_from_buckets(self) -> None:
-        """Total cluster count equals sum of bucket counts."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("15min"), "99", "2000", "198000", -1),
-            make_trade(now + pd.Timedelta("20min"), "101", "3000", "303000", 1),
-            make_trade(now + pd.Timedelta("25min"), "98", "4000", "392000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 1)
-        total = sum(data[0][f"p{p}"]["totalClusterCount"] for p in [0, 25, 50, 75, 90, 95, 99])
-        self.assertEqual(total, 4)
-
-    def test_buy_sell_in_buckets(self) -> None:
-        """Buy/sell breakdown correct in percentile buckets."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("15min"), "99", "2000", "198000", -1),
-            make_trade(now + pd.Timedelta("20min"), "101", "3000", "303000", 1),
-            make_trade(now + pd.Timedelta("25min"), "98", "4000", "392000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 1)
-        total_buy = sum(data[0][f"p{p}"]["buyClusterCount"] for p in [0, 25, 50, 75, 90, 95, 99])
-        total_sell = sum(
-            data[0][f"p{p}"]["totalClusterCount"] - data[0][f"p{p}"]["buyClusterCount"]
-            for p in [0, 25, 50, 75, 90, 95, 99]
-        )
-        self.assertEqual(total_buy, 2)
-        self.assertEqual(total_sell, 2)
-
-    def test_notional_in_buckets(self) -> None:
-        """Notional values present in buckets."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("20min"), "99", "2000", "198000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 1)
-        total_notional = sum(
-            data[0][f"p{p}"]["totalClusterNotional"] for p in [0, 25, 50, 75, 90, 95, 99]
-        )
-        self.assertEqual(total_notional, 298000.0)
-
-    def test_empty_window(self) -> None:
-        """Window with no trades handled."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("90min"), "100", "1000", "100000", 1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("2h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 1)
-        self.assertEqual(data[0]["timestamp"], now + pd.Timedelta("1h"))
-
-    def test_ohlcv_fields_present(self) -> None:
-        """Standard OHLCV fields present."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("20min"), "105", "1000", "105000", 1),
-            make_trade(now + pd.Timedelta("30min"), "95", "1000", "95000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 1)
-        self.assertIn("open", data[0])
-        self.assertIn("high", data[0])
-        self.assertIn("low", data[0])
-        self.assertIn("close", data[0])
-        self.assertIn("volume", data[0])
-
-    def test_daily_window(self) -> None:
-        """Daily window works correctly."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1d")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("6h"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("30h"), "105", "1000", "105000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("2d"), df, warmed_cache())
-
-        self.assertEqual(len(data), 2)
-        self.assertEqual(data[0]["timestamp"], now)
-        self.assertEqual(data[1]["timestamp"], now + pd.Timedelta("1d"))
-
-    def test_clusters_separate_across_windows(self) -> None:
-        """Clusters in different windows stay separate."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("30min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("90min"), "101", "2000", "202000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("2h"), df, warmed_cache())
-
-        self.assertEqual(len(data), 2)
-        self.assertEqual(float(data[0]["volume"]), 1000.0)
-        self.assertEqual(float(data[1]["volume"]), 2000.0)
-
-    def test_partial_cluster_merged_same_window(self) -> None:
-        """Partial cluster merged with next batch if same window and direction."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1d")
-
-        df1 = pd.DataFrame([
-            make_trade(now + pd.Timedelta("1h"), "100", "1000", "100000", 1),
-        ])
-        data1, cache = candle.aggregate(now, now + pd.Timedelta("2h"), df1, warmed_cache())
-        self.assertEqual(len(data1), 0)
-        self.assertIn("partial_cluster", cache)
-
-        df2 = pd.DataFrame([
-            make_trade(now + pd.Timedelta("3h"), "101", "2000", "202000", 1),
-        ])
-        data2, cache2 = candle.aggregate(
-            now + pd.Timedelta("2h"), now + pd.Timedelta("4h"), df2, cache
-        )
-
-        self.assertEqual(len(data2), 0)
-        self.assertIn("partial_cluster", cache2)
-        self.assertEqual(float(cache2["partial_cluster"]["volume"]), 3000.0)
-
-    def test_partial_cluster_flushed_at_window_boundary(self) -> None:
-        """Partial cluster flushed at window boundary, not merged."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1d")
-
-        df1 = pd.DataFrame([
-            make_trade(now + pd.Timedelta("23h"), "100", "1000", "100000", 1),
-        ])
-        data1, cache = candle.aggregate(now, now + pd.Timedelta("23h30min"), df1, warmed_cache())
-        self.assertEqual(len(data1), 0)
-        self.assertIn("partial_cluster", cache)
-
-        df2 = pd.DataFrame([
-            make_trade(now + pd.Timedelta("25h"), "101", "2000", "202000", 1),
-        ])
-        data2, cache2 = candle.aggregate(
-            now + pd.Timedelta("23h30min"), now + pd.Timedelta("26h"), df2, cache
-        )
-
-        self.assertEqual(len(data2), 1)
-        self.assertEqual(float(data2[0]["volume"]), 1000.0)
-        self.assertIn("partial_cluster", cache2)
-        self.assertEqual(float(cache2["partial_cluster"]["volume"]), 2000.0)
-
-    def test_cross_batch_stats_accumulate(self) -> None:
-        """Stats accumulate correctly across batches in same window."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle(window="1h")
-
-        df1 = pd.DataFrame([
-            make_trade(now + pd.Timedelta("5min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("10min"), "100", "2000", "200000", -1),
-        ])
-        data1, cache1 = candle.aggregate(now, now + pd.Timedelta("15min"), df1, warmed_cache())
-
-        self.assertEqual(len(data1), 0)
-        self.assertIn("next", cache1)
-
-        df2 = pd.DataFrame([
-            make_trade(now + pd.Timedelta("45min"), "100", "3000", "300000", 1),
-            make_trade(now + pd.Timedelta("55min"), "100", "4000", "400000", -1),
-        ])
-        data2, cache2 = candle.aggregate(
-            now + pd.Timedelta("15min"), now + pd.Timedelta("1h5min"), df2, cache1
-        )
-
-        self.assertEqual(len(data2), 1)
-        result = data2[0]
-
-        self.assertEqual(float(result["volume"]), 10000.0)
-
-        total_clusters = sum(
-            result[f"p{p}"]["totalClusterCount"]
-            for p in [0, 25, 50, 75, 90, 95, 99]
-        )
-        self.assertEqual(total_clusters, 4)
-
-    def test_no_bucket_stats_when_disabled(self) -> None:
-        """No bucket stats when cluster_bucket_stats is false."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = TimeBasedCandle(json_data={
-            "source_data": FileData.FILTERED,
-            "window": "1h",
-            "cluster_bucket_stats": False,
-        })
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, {})
-
-        self.assertEqual(len(data), 1)
-        self.assertNotIn("p50", data[0])
-        self.assertIn("open", data[0])
-
-
-class ConstantClusterBucketTest(TestCase):
-    """Test ClusterBucketMixin with ConstantCandle."""
-
-    def _make_candle(self, **overrides: object) -> ConstantCandle:
-        """Create constant candle with cluster bucket stats enabled."""
-        json_data = {
-            "source_data": FileData.FILTERED,
-            "sample_type": "notional",
-            "target_value": Decimal("500000"),
-            "cluster_bucket_stats": True,
-        }
-        json_data.update(overrides)
-        return ConstantCandle(json_data=json_data)
-
-    def test_bucket_stats_present_in_constant_candle(self) -> None:
-        """Bucket stats present when candle completes."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle()
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("1min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("2min"), "100", "2000", "200000", -1),
-            make_trade(now + pd.Timedelta("3min"), "100", "3000", "300000", 1),
-        ])
-
-        cache = warmed_cache()
-        cache["sample_value"] = 0
-        data, cache = candle.aggregate(now, now + pd.Timedelta("5min"), df, cache)
-
-        self.assertEqual(len(data), 1)
-        for p in [0, 25, 50, 75, 90, 95, 99]:
-            key = f"p{p}"
-            self.assertIn(key, data[0])
-
-    def test_cluster_count_matches_trades(self) -> None:
-        """Total clusters matches number of direction changes."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = self._make_candle()
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("1min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("2min"), "100", "2000", "200000", -1),
-            make_trade(now + pd.Timedelta("3min"), "100", "3000", "300000", 1),
-        ])
-
-        cache = warmed_cache()
-        cache["sample_value"] = 0
-        data, cache = candle.aggregate(now, now + pd.Timedelta("5min"), df, cache)
-
-        self.assertEqual(len(data), 1)
-        total = sum(data[0][f"p{p}"]["totalClusterCount"] for p in [0, 25, 50, 75, 90, 95, 99])
-        self.assertEqual(total, 3)
-
-
-class ClusterEmaWarmupTest(TestCase):
-    """Test EMA warmup behavior - no candles emitted until EMA is ready."""
-
-    def test_no_candles_during_warmup_time_based(self) -> None:
-        """TimeBasedCandle emits no candles during EMA warmup."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = TimeBasedCandle(json_data={
-            "source_data": FileData.FILTERED,
-            "window": "1h",
-            "cluster_bucket_stats": True,
-        })
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("90min"), "101", "1000", "101000", -1),
-        ])
-
-        # Empty cache - EMA count is 0
-        data, cache = candle.aggregate(now, now + pd.Timedelta("2h"), df, {})
-
-        # No candles emitted during warmup
-        self.assertEqual(len(data), 0)
-        # But EMA is being updated
-        self.assertIn("cluster_ema_count", cache)
-        self.assertGreater(cache["cluster_ema_count"], 0)
-
-    def test_no_candles_during_warmup_constant(self) -> None:
-        """ConstantCandle emits no candles during EMA warmup."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = ConstantCandle(json_data={
-            "source_data": FileData.FILTERED,
-            "sample_type": "notional",
-            "target_value": Decimal("500000"),
-            "cluster_bucket_stats": True,
-        })
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("1min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("2min"), "100", "2000", "200000", -1),
-            make_trade(now + pd.Timedelta("3min"), "100", "3000", "300000", 1),
-        ])
-
-        # Empty cache - EMA count is 0
-        data, cache = candle.aggregate(now, now + pd.Timedelta("5min"), df, {"sample_value": 0})
-
-        # No candles emitted during warmup
-        self.assertEqual(len(data), 0)
-        # But EMA is being updated
-        self.assertIn("cluster_ema_count", cache)
-        self.assertGreater(cache["cluster_ema_count"], 0)
-
-    def test_ema_accumulates_during_warmup(self) -> None:
-        """EMA continues to accumulate during warmup."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = TimeBasedCandle(json_data={
-            "source_data": FileData.FILTERED,
-            "window": "1min",
-            "cluster_bucket_stats": True,
-        })
-
-        # Process multiple batches
-        cache = {}
-        for i in range(10):
-            ts = now + pd.Timedelta(f"{i * 10}min")
-            df = pd.DataFrame([
-                make_trade(ts + pd.Timedelta("30s"), "100", str(1000 + i * 100), "100000", 1),
-            ])
-            data, cache = candle.aggregate(ts, ts + pd.Timedelta("10min"), df, cache)
-
-        # No candles emitted (still in warmup)
-        self.assertEqual(len(data), 0)
-        # But EMA count accumulated
-        self.assertEqual(cache["cluster_ema_count"], 10)
-
-    def test_candles_emitted_after_warmup(self) -> None:
-        """Candles start being emitted once EMA is warmed up."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = TimeBasedCandle(json_data={
-            "source_data": FileData.FILTERED,
-            "window": "1h",
-            "cluster_bucket_stats": True,
-        })
-
-        # Start with almost-warmed cache (1 cluster away from ready)
-        cache = {
-            "cluster_ema_count": WARMUP_MIN_CLUSTERS - 1,
-            "cluster_ema_mean": 7.0,
-            "cluster_ema_var": 1.0,
-        }
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-            make_trade(now + pd.Timedelta("90min"), "101", "1000", "101000", -1),
-        ])
-
-        data, cache = candle.aggregate(now, now + pd.Timedelta("2h"), df, cache)
-
-        # Both windows emit - first reaches WARMUP_MIN_CLUSTERS and emits immediately
-        self.assertEqual(len(data), 2)
-        self.assertEqual(data[0]["timestamp"], now)
-        self.assertEqual(data[1]["timestamp"], now + pd.Timedelta("1h"))
-
-    def test_warmup_disabled_when_cluster_stats_disabled(self) -> None:
-        """No warmup delay when cluster_bucket_stats is disabled."""
-        now = pd.Timestamp("2024-01-01 00:00:00", tz="UTC")
-        candle = TimeBasedCandle(json_data={
-            "source_data": FileData.FILTERED,
-            "window": "1h",
-            "cluster_bucket_stats": False,
-        })
-
-        df = pd.DataFrame([
-            make_trade(now + pd.Timedelta("10min"), "100", "1000", "100000", 1),
-        ])
-
-        # Empty cache
-        data, cache = candle.aggregate(now, now + pd.Timedelta("1h"), df, {})
-
-        # Candle emitted immediately
-        self.assertEqual(len(data), 1)
