+++ quant_tick/views/__init__.py
@@ -1,12 +1,10 @@
 from .aggregate_candles import AggregateCandleView
 from .aggregate_trades import AggregateTradeDataView
 from .candles import CandleDataView, CandleView
-from .inference import InferenceView
 
 __all__ = [
     "AggregateCandleView",
     "AggregateTradeDataView",
     "CandleDataView",
     "CandleView",
-    "InferenceView",
 ]
